MONIT-5441: Refactored the IngesterFormatter code.
