+++ java-lib/src/main/java/com/wavefront/ingester/GraphiteDecoder.java
@@ -20,7 +20,8 @@ import sunnylabs.report.ReportPoint;
 public class GraphiteDecoder implements Decoder<String> {
 
   private static final Pattern CUSTOMERID = Pattern.compile("[a-z]+");
-  private static final IngesterFormatter FORMAT = IngesterFormatter.newBuilder().whiteSpace()
+  private static final IngesterFormatter FORMAT = (IngesterFormatter) IngesterFormatter.newBuilder()
+      .whiteSpace()
       .appendMetricName().whiteSpace()
       .appendValue().whiteSpace()
       .appendOptionalTimestamp().whiteSpace()
