+++ java-lib/src/main/java/com/wavefront/ingester/OpenTSDBDecoder.java
@@ -16,7 +16,8 @@ import sunnylabs.report.ReportPoint;
 public class OpenTSDBDecoder implements Decoder<String> {
 
   private final String hostName;
-  private static final IngesterFormatter FORMAT = IngesterFormatter.newBuilder().whiteSpace()
+  private static final IngesterFormatter FORMAT = (IngesterFormatter) IngesterFormatter.newBuilder()
+      .whiteSpace()
       .appendCaseInsensitiveLiteral("put").whiteSpace()
       .appendMetricName().whiteSpace()
       .appendTimestamp().whiteSpace()
