+++ proxy/src/test/java/com/wavefront/agent/SourceTagHandlerTest.java
@@ -1,7 +1,6 @@
 package com.wavefront.agent;
 
 import com.wavefront.agent.api.ForceQueueEnabledAgentAPI;
-import com.wavefront.api.AgentAPI;
 
 import org.easymock.EasyMock;
 import org.junit.BeforeClass;
@@ -9,7 +8,6 @@ import org.junit.Test;
 
 import java.io.IOException;
 import java.util.Arrays;
-import java.util.List;
 import java.util.UUID;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
